同步源

+++ dp2Circulation/OperLog/OperLogForm.cs
@@ -15,6 +15,8 @@ using System.Threading;
 using System.IO;
 using System.Web;
 
+using ClosedXML.Excel;
+
 using DigitalPlatform;
 using DigitalPlatform.GUI;
 using DigitalPlatform.Xml;
@@ -23,13 +25,10 @@ using DigitalPlatform.Range;
 using DigitalPlatform.Text;
 using DigitalPlatform.Marc;
 using DigitalPlatform.MarcDom;
-
-// using DigitalPlatform.LibraryClient.localhost;
 using DigitalPlatform.CommonControl;
 using DigitalPlatform.CirculationClient;
 using DigitalPlatform.LibraryClient;
 using DigitalPlatform.LibraryClient.localhost;
-using ClosedXML.Excel;
 
 namespace dp2Circulation
 {
@@ -2127,6 +2126,8 @@ DomUtil.GetElementInnerXml(dom.DocumentElement, "deletedCommentRecords"));
             {
                 if (strAction == "createDatabase")
                     return "创建数据库";
+                if (strAction == "changeDatabase")
+                    return "修改数据库定义";
                 if (strAction == "deleteDatabase")
                     return "删除数据库";
                 if (strAction == "initializeDatabase")
@@ -3930,6 +3931,11 @@ FileShare.ReadWrite))
                 }
             }
 
+            if (strOperation == "crashReport")
+            {
+                string strSubject = DomUtil.GetElementText(dom.DocumentElement, "subject");
+                param.Comment = strSubject;
+            }
             return true;
 
         ERROR1:
@@ -7029,6 +7035,7 @@ Keys keyData)
         /// <param name="filenames">要参与处理的日志文件名集合</param>
         /// <param name="nLevel">从 dp2Library 服务器获取日志记录的详细级别</param>
         /// <param name="strStyle">处理风格。autocache</param>
+        /// <param name="strFilter">过滤参数</param>
         /// <param name="strCacheDir">日志本地缓存目录</param>
         /// <param name="param">回调对象</param>
         /// <param name="procDoRecord">回调函数</param>
